// src/app/products/product-category.enum.ts
export enum ProductCategory {
  ACESSORIOS = 'ACESSORIOS',
  ANEIS = 'ANEIS',
  BERLOQUES = 'BERLOQUES',
  BRINCOS = 'BRINCOS',
  CAMISETAS = 'CAMISETAS',
  COLARES = 'COLARES',
  CONJUNTOS = 'CONJUNTOS',
  DECORACAO = 'DECORACAO',
  PULSEIRAS = 'PULSEIRAS',
  MAE_EMPREENDEDORA = 'MAE_EMPREENDEDORA',
  MAE_LOUCA_POR_FRIDA = 'MAE_LOUCA_POR_FRIDA',
  MAE_CASEIRA = 'MAE_CASEIRA',
  MAE_DESCOLADA = 'MAE_DESCOLADA',
  MAE_CLASSICA = 'MAE_CLÁSSICA',
}